Adds support for creating a project with visibility
